refactor: create ambient_project_rt
