+++ guest/rust/Cargo.toml
@@ -9,6 +9,7 @@ rust-version = "1.67.0"
 ambient_api_macros = { path = "api/api_macros", version = "0.2.0-dev" }
 ambient_api = { path = "api", version = "0.2.0-dev" }
 
+ambient_project_rt = { path = "../../shared_crates/project_rt", version = "0.2.0-dev" }
 ambient_shared_types = { path = "../../shared_crates/shared_types", version = "0.2.0-dev" }
 
 ambient_project = { path = "../../shared_crates/project", default-features = false, features = ["guest"], version = "0.2.0-dev" }
