+++ shared_crates/project/Cargo.toml
@@ -14,8 +14,6 @@ proc-macro2 = { version = "1.0", optional = true }
 serde = { workspace = true }
 toml = { workspace = true }
 thiserror = { workspace = true }
-byteorder = { workspace = true }
-glam = { workspace = true }
 
 [dev-dependencies]
 paste = { workspace = true }
