+++ shared_crates/project_rt/README.md
@@ -1,0 +1,3 @@
+# Ambient project runtime
+
+Runtime support for code generated by the Ambient project macro.
