+++ shared_crates/project_rt/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod message_serde;
